apply clippy

+++ src/grid.rs
@@ -1,9 +1,11 @@
 use crate::components::{Position, Size};
 use bevy::prelude::*;
 
-const GRID_WIDTH: u32 = 10;
-const GRID_HEIGHT: u32 = 10;
+const GRID_WIDTH: u16 = 10;
+const GRID_HEIGHT: u16 = 10;
 
+#[allow(clippy::missing_panics_doc)]
+#[allow(clippy::needless_pass_by_value)]
 pub fn size_scaling(windows: Res<Windows>, mut q: Query<(&Size, &mut Transform)>) {
     let window = windows.get_primary().unwrap();
     for (sprite_size, mut transform) in q.iter_mut() {
@@ -11,6 +13,8 @@ pub fn size_scaling(windows: Res<Windows>, mut q: Query<(&Size, &mut Transform)>
     }
 }
 
+#[allow(clippy::missing_panics_doc)]
+#[allow(clippy::needless_pass_by_value)]
 pub fn position_translation(windows: Res<Windows>, mut q: Query<(&Position, &mut Transform)>) {
     let window = windows.get_primary().unwrap();
     for (pos, mut transform) in q.iter_mut() {
@@ -18,6 +22,7 @@ pub fn position_translation(windows: Res<Windows>, mut q: Query<(&Position, &mut
     }
 }
 
+#[allow(clippy::cast_lossless)]
 fn scale_sprite(transform: &mut Transform, sprite_size: &Size, window: &Window) {
     transform.scale = Vec3::new(
         sprite_size.width / GRID_WIDTH as f32 * window.width() as f32,
@@ -26,11 +31,13 @@ fn scale_sprite(transform: &mut Transform, sprite_size: &Size, window: &Window) 
     );
 }
 
+#[allow(clippy::cast_lossless)]
 fn convert(pos: f32, bound_window: f32, grid_side_lenght: f32) -> f32 {
     let tile_size = bound_window / grid_side_lenght;
     pos / grid_side_lenght * bound_window - (bound_window / 2.) + (tile_size / 2.)
 }
 
+#[allow(clippy::cast_lossless)]
 fn translate_position(transform: &mut Transform, pos: &Position, window: &Window) {
     transform.translation = Vec3::new(
         convert(pos.x as f32, window.width() as f32, GRID_WIDTH as f32),
